final changes for pmax 4.4-Lite II release
SCCS-vsn: 8.5

+++ sys/pmax/pmax/pmap.c
@@ -34,7 +34,7 @@
  * OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF
  * SUCH DAMAGE.
  *
- *	@(#)pmap.c	8.4 (Berkeley) 01/26/94
+ *	@(#)pmap.c	8.5 (Berkeley) 06/02/95
  */
 
 /*
@@ -1439,7 +1439,7 @@ vm_page_alloc1()
 	if (cnt.v_free_count < cnt.v_free_min ||
 	    (cnt.v_free_count < cnt.v_free_target &&
 	     cnt.v_inactive_count < cnt.v_inactive_target))
-		thread_wakeup((int)&vm_pages_needed);
+		thread_wakeup(&vm_pages_needed);
 	return (mem);
 }
